# opt: Improve list and optbuilder perf