- Improve perf of list storage
- Improve perf of USING and natural joins

Together, the changes give a perf boost for this case:
```
SELECT * FROM foo NATURAL JOIN bar

name                 old time/op    new time/op    delta
Phases/test/Explore     115µs ± 2%     109µs ± 1%  -4.63%  (p=0.008 n=5+5)

name                 old alloc/op   new alloc/op   delta
Phases/test/Explore    44.3kB ± 0%    42.6kB ± 0%  -3.76%  (p=0.008 n=5+5)

name                 old allocs/op  new allocs/op  delta
Phases/test/Explore       189 ± 0%       181 ± 0%  -4.23%  (p=0.008 n=5+5)
```